updated repository folder added player repository test
